Added support to read joblist and node list from configfiles. There is a now a ste simulation in the admin adgent that check if it is time to add a new job or node into the simulatio.
TODO: change simulation step in general simulation agent,
as for now it's not possible to update the mobile node status
while running a job because the simulator performs a sleep command
that stop everithing. We need to change this behavior to have
a simulation step than advance the simulation every second
using a periodica besa guard.